debloat the bridge and refactor out components
